requirements.txt
# Streamlit App
streamlit==1.52.0
python-dotenv==1.0.1

# LangChain & OpenAI (Atualizadas para 2025 - Fixed OpenAI version)
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
openai>=1.108.1,<2.0.0

# Anthropic (para Claude models)
langchain-anthropic==0.3.0
anthropic>=0.40.0

# LangGraph (Orquestração com State Machines)
langgraph==0.2.62
langgraph-checkpoint==2.0.8

# LlamaIndex & Parse
llama-parse==0.6.83
llama-cloud==0.1.45
llama-index-core==0.14.9
llama-index-embeddings-openai==0.5.1
llama-index-llms-openai==0.6.10

# Retry & Resilience
tenacity==9.1.2

# Data Processing
python-docx==1.1.0

# Numeric
numpy>=1.26.0

# Vector Store
chromadb>=0.5.0

# Utilities
pydantic==2.12.5
